Remove accidentally left extra thread in can-message.lisp
Also added a catch all pattern to the function receiving thread to
detect if incorrect messages are sent. This was copied from (I think?)
bat-ant-esp main.lisp from some branch as reference. For some reason
having this thread present seems to cause lbm on jet-if-esp to crash
after exactly 94 sends. ¯\_(ツ)_/¯