Add outlier detection to jet temperature sampling
For each sample, we check that the sample is larger than the lowest
25 %, but also lower then the highest 75 % of the last 40 samples
(roughly the last 40 ms), and if it passes, we pass it along to the
previous low pass filter.
I.e. if we devide the last 40 samples (as a sorted list) into four
equally large partitions, we only use a sample if its in the second or
thrird partitions (Q2 or Q3).