Kind of worthless magn value experiment
Tried playing around with selecting the magnometer values with the
greatest change. Came to the conclusion that it doens't matter, so
we're still using all of the values. But I'm keeping the code around
because I'm an irresponsible code horder (and to be able to play
around with it in the future).